backend/requirements.txt
# Core
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3

# Auth & Security
djangorestframework-simplejwt>=5.4
django-environ>=0.11
python-decouple>=3.8

# DB & ORM
psycopg[binary]>=3.2
dj-database-url>=2.2

# Async Tasks / Notifications
celery>=5.4
redis>=5.0  # broker for celery

# Validation
pydantic[email]>=2.7

# File & Storage
django-storages>=1.14  # if you later integrate S3/GCP/Azure
Pillow>=10.0           # handle images

# API Extras
django-ninja>=1.2  # optional, for schema-driven APIs
django-ninja-jwt[crypto]>=5.3  # optional, JWT with ninja

# Developer Tools
black>=24.0
isort>=5.13
pytest>=8.2
pytest-django>=4.9
